worker: enable ClickHouse logging support
This enables support for logging requests/responses for Nix cache
objects, through log insertion into ClickHouse tables. Data is ingested
using the ClickHouse HTTP frontend, configured with HTTP Basic
Authentication, over HTTPS, at the specified ingestion domain. There is
no ClickHouse proxy or frontend at the moment: this is just a
single-node server that only allows a single user, the CloudFlare
worker, to insert data, over a public HTTP address.

The underlying data is inserted as JSON, using the JSONEachRow format in
order to match columns to data objects. This gives an extremely simple
and clean mapping between the worker object and the underlying
ClickHouse schema.

The log POST request must be handled in the worker as an asynchronous
event that *must* be completed for accurate logs. However, simply
awaiting the result incurs a high extra latency cost for every request,
since we wait for the ClickHouse server before returning a response to
the client. Luckily, this is easy to fix: the usage of event.waitUntil
allows us to register a Promise with the Worker runtime that will allow
us to return a response to the client immediately, while ensuring the
POST completes before worker shutdown.

Signed-off-by: Austin Seipp <aseipp@pobox.com>